+++ src/main/shared/types/serverConfig.ts
@@ -1,0 +1,9 @@
+export interface ServerConfig {
+    hideHealthArmour?: boolean;
+    hideMinimapOnFoot?: boolean;
+    hideMinimapInPage?: boolean;
+    hideVehicleName?: boolean;
+    hideVehicleClass?: boolean;
+    hideStreetName?: boolean;
+    hideAreaName?: boolean;
+}